refactor: split app.py into ui/ package
Extract the 784-line Streamlit monolith into six focused modules:
- ui/data_loaders.py  — @st.cache_data wrappers
- ui/term_actions.py  — define/explain LLM logic
- ui/sidebar.py       — sidebar rendering
- ui/metadata_panel.py — left-column analysis expanders
- ui/transcript_browser.py — searchable HTML transcript component
- ui/feynman.py       — Feynman loop + chat interface

app.py is now a ~100-line orchestrator. Closes #26.
